Creadas las tablas y rellaciones ER

+++ database/migrations/mios/2019_04_12_010940_crear_tabla_usuario_rol.php
@@ -14,10 +14,10 @@ class CrearTablaUsuarioRol extends Migration
     public function up()
     {
         Schema::create('usuario_rol', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('rol_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('rol_id');
             $table->foreign('rol_id', 'fk_usuariorol_rol')->references('id')->on('rol')->onDelete('restrict')->onUpdate('restrict');
-            $table->unsignedInteger('usuario_id');
+            $table->unsignedBigInteger('usuario_id');
             $table->foreign('usuario_id', 'fk_usuariorol_usuario')->references('id')->on('usuario')->onDelete('restrict')->onUpdate('restrict');
             $table->boolean('estado');
             $table->timestamps();
